codex: address PR review feedback (#2330)
