+++ apps/aevatar-console-web/src/shared/navigation/teamRoutes.ts
@@ -224,6 +224,37 @@ export function buildTeamMemberInvokeHref(options: {
   return `/scopes/${encodeURIComponent(scopeId)}/teams/${encodeURIComponent(teamId)}/members/${encodeURIComponent(memberId)}/invoke`;
 }
 
+export function buildTeamMemberPublishedRunsHref(options: {
+  actorId?: string;
+  memberId?: string;
+  runId?: string;
+  scopeId: string;
+  teamId: string;
+}): string {
+  const scopeId = trimOptional(options.scopeId);
+  const teamId = trimOptional(options.teamId);
+  if (!scopeId || !teamId) {
+    return buildTeamsHref();
+  }
+
+  const memberId = trimOptional(options.memberId);
+  if (!memberId) {
+    return buildTeamDetailHref({
+      scopeId,
+      tab: 'members',
+      teamId,
+    });
+  }
+
+  return buildHref(
+    `/scopes/${encodeURIComponent(scopeId)}/teams/${encodeURIComponent(teamId)}/members/${encodeURIComponent(memberId)}/runs`,
+    {
+      runId: options.runId,
+      actorId: options.actorId,
+    },
+  );
+}
+
 export function buildTeamMemberAutomationsHref(options: {
   memberId?: string;
   scopeId: string;
